feat: renaming prototype folder
