+++ src/js/synth/Analyser.js
