+++ src/js/synth/Chorus.js
