+++ src/js/synth/Operator.js
