+++ src/js/synth/Voice.js
